pyproject.toml
[project]
name = "comp-analyst-agent"
version = "1.0.0"
description = "Take-home exercise: Build a compensation analyst agent"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]
